refactor: isolate user creation logic into InitialUserDataService and update CreateNewUser
